move crosshair changes to baseplot from timeplot
