+++ pydm/widgets/timeplot.py
@@ -2,10 +2,9 @@ import time
 import json
 from collections import OrderedDict
 from typing import Optional
-from pyqtgraph import BarGraphItem, ViewBox, AxisItem, PlotDataItem, TextItem, mkBrush, mkPen
+from pyqtgraph import BarGraphItem, ViewBox, AxisItem
 import numpy as np
-from datetime import datetime
-from qtpy.QtGui import QColor, QFont, QCursor
+from qtpy.QtGui import QColor, QCursor
 from qtpy.QtCore import Signal, Slot, Property, QTimer, QPointF
 from .baseplot import BasePlot, BasePlotCurveItem
 from .channel import PyDMChannel
@@ -561,10 +560,6 @@ class PyDMTimePlot(BasePlot):
 
         self.auto_scroll_timer = QTimer()
         self.auto_scroll_timer.timeout.connect(self.auto_scroll)
-
-        self.textItems = {}
-        self.crosshair = False
-        self.init_labels = False
 
     def to_dict(self) -> OrderedDict:
         """Adds attribute specific to TimePlot to add onto BasePlot's to_dict.
@@ -1206,127 +1201,6 @@ class PyDMTimePlot(BasePlot):
             vertical_movable,
             horizontal_movable,
         )
-
-        if is_enabled:
-            self.crosshair = True
-            self.textItems = {}
-            self.init_label = True
-            self.crosshair_position_updated.connect(self.updateLabel)
-        else:
-            self.clearCurveLabels()
-
-    def initializeCurveLabels(self, font: str = "arial", font_size: int = 8) -> None:
-        """
-        Create a TextItem for each PlotDataItem in the plot and stores them in self.textItems.
-
-        Returns
-        -------
-        None
-        """
-        self.clearCurveLabels()
-        self.init_label = False
-
-        for item in self.plotItem.listDataItems():
-            if not isinstance(item, PlotDataItem):
-                continue
-
-            label: TextItem = TextItem(
-                text="No data", color="w", border=mkPen(color="w", width=2), fill=mkBrush(0, 0, 0, 150)
-            )
-
-            label.setPos(0, 0)
-            label.setAnchor((0.5, 0.5))
-            label.setFont(QFont(font, font_size))
-
-            self.textItems[item] = label
-
-    def clearCurveLabels(self) -> None:
-        """
-        Remove all existing curve labels from the plot and clear the textItems dictionary.
-
-        This method iterates through all TextItem objects stored in the `textItems` attribute,
-        removes each one from the plot (via `plotItem.removeItem`), clears the dictionary, and
-        sets the `init_label` flag to True.
-
-        Returns
-        -------
-        None
-        """
-        if hasattr(self, "textItems"):
-            for label in self.textItems.values():
-                self.plotItem.removeItem(label)
-            self.textItems.clear()
-            self.init_label = True
-
-    @Slot(float, float)
-    def updateLabel(self, x_val: float, y_val: float) -> None:
-        """
-        Update the label for each curve based on the given x-coordinate.
-
-        For each curve stored in the `textItems` dictionary, this method retrieves the curve's
-        data (x and y arrays) and finds the data point with an x-value immediately to the left
-        of `x_val`. If the x-coordinate is within the range of the curve's data and the data point
-        is finite, the corresponding label is updated to display the x and y values and is moved
-        to that position. If `x_val` is not finite or is outside the data range, the label text is
-        set to "No data!".
-
-        Parameters
-        ----------
-        x_val : float
-            The x-coordinate (in data space) used to determine which data point to label.
-
-        Returns
-        -------
-        None
-        """
-        if not np.isfinite(x_val):
-            return
-
-        if self.init_label:
-            self.initializeCurveLabels()
-            self.init_label = False
-
-        for curve, label in self.textItems.items():
-            xData, yData = curve.getData()
-            if xData is None or yData is None or len(xData) == 0:
-                label.setText("No data!")
-                continue
-
-            if x_val < xData[0] or x_val > xData[-1]:
-                label.setText("No data!")
-                continue
-
-            idx = np.searchsorted(xData, x_val, side="right") - 1
-            if idx < 0:
-                idx = 0
-            if idx >= len(yData):
-                idx = len(yData) - 1
-
-            real_x = xData[idx]
-            real_y = yData[idx]
-
-            if not (np.isfinite(real_x) and np.isfinite(real_y)):
-                continue
-
-            if hasattr(curve, "y_axis_name") and curve.y_axis_name in self.plotItem.axes:
-                curve_vb = self.plotItem.getViewBoxForAxis(curve.y_axis_name)
-            else:
-                curve_vb = self.getViewBox()
-
-            curve_vb.addItem(label)
-
-            time_str = datetime.fromtimestamp(real_x).strftime("%H:%M:%S")
-
-            if curve.severity_raw != -1:
-                label.setText(f"x={time_str}\ny={real_y:.2f}" + "\n" + str(curve.severity))
-            else:
-                label.setText(f"x={time_str}\ny={real_y:.2f}")
-
-            data_point = QPointF(x_val, y_val)
-            primary_vb = self.getViewBox()
-            scene_point = primary_vb.mapViewToScene(data_point)
-            label_point = curve_vb.mapSceneToView(scene_point)
-            label.setPos(label_point.x(), real_y)
 
 
 class TimeAxisItem(AxisItem):
